Split FormInput in multiple components for easier and clearer control of the form
